Relocated Affinity Patches
and added something else but trust me it is not that important guys please dont look
